Components: move filterBox to components folder
